fixup! refactor(devtools): disable unsupported features
