+++ devtools/projects/ng-devtools-backend/src/lib/component-tree/core-enums.ts
